requirements.txt
pandas 
numpy 
Flask
matplotlib
seaborn
python-dotenv
pymongo
pymongo
certifi
scikit-learn
dotenv
pyaml
mlflow
# -e .
